Moved indice and vertex generation to `render.renderer.render_data_util`
